+++ src/extension/utilities/DeclarationIndexMapper.ts
